refactor: reorganize imports and move mapPluginGuardToAccountGuard to a new file
