+++ packages/instance-registry/src/service-keycloak-execution.ts
@@ -2,7 +2,8 @@ import { createSdkLogger } from '@sva/server-runtime';
 import type { InstanceKeycloakProvisioningRun } from '@sva/core';
 import type { ExecuteInstanceKeycloakProvisioningInput } from './mutation-types.js';
 import type { InstanceRegistryServiceDeps } from './service-types.js';
-import { createGetKeycloakStatusHandler, loadInstanceWithSecret } from './service-keycloak.js';
+import { createGetKeycloakStatusHandler } from './service-keycloak-readers.js';
+import { loadInstanceWithSecret } from './service-keycloak-secrets.js';
 import { appendRunStep } from './service-keycloak-run-steps.js';
 import { buildProvisioningInput, completeRun, createQueuedRun, readQueuedTemporaryPassword, syncProvisionedClientSecretToRegistry, syncRotatedClientSecretToRegistry } from './service-keycloak-execution-shared.js';
 import { failClaimedRun, failRun } from './service-keycloak-execution-failures.js';
